server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Andrew Gordon"]

[tool.poetry.dependencies]
python = "^3.6"
djangorestframework = "^3.10"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
django = "^2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
